refactor: 숫자가 아닌 입력 값에 대한 유효성 검증을 LottoNumber내에서 처리하도록 수정

+++ src/main/kotlin/controller/LottoController.kt
@@ -51,7 +51,7 @@ class LottoController(private val inputView: InputView, private val outputView: 
     private fun generateWinningNumbers(): Pair<Lotto, LottoNumber> {
         outputView.printWinningNumbersMessage()
         val winningNumbers =
-            Lotto(LottoNumbers(inputView.readWinningNumbers().map { LottoNumber(it.toInt()) }))
+            Lotto(LottoNumbers(inputView.readWinningNumbers().map { LottoNumber(it) }))
         outputView.printBonusNumberMessage()
         val bonusNumber = LottoNumber(inputView.readWinningBonusNumber())
         return Pair(winningNumbers, bonusNumber)
